Split exec::view mod
